Avoid compilation warning
Comparison of uint64_t and int64_t causes compilation warning!

+++ EventFiltering/PWGUD/diffractionFilter.cxx
@@ -58,7 +58,7 @@ T compatibleBCs(soa::Join<aod::Collisions, aod::EvSels>::iterator const& collisi
   uint64_t meanBC = mostProbableBC - std::lround(collision.collisionTime() / o2::constants::lhc::LHCBunchSpacingNS);
   int deltaBC = std::ceil(collision.collisionTimeRes() / o2::constants::lhc::LHCBunchSpacingNS * ndt);
   int64_t minBC = meanBC - deltaBC;
-  int64_t maxBC = meanBC + deltaBC;
+  uint64_t maxBC = meanBC + deltaBC;
   if (minBC < 0) {
     minBC = 0;
   }
@@ -66,7 +66,7 @@ T compatibleBCs(soa::Join<aod::Collisions, aod::EvSels>::iterator const& collisi
   // find slice of BCs table with BC in [minBC, maxBC]
   int64_t maxBCId = bcIter.globalIndex();
   int moveCount = 0; // optimize to avoid to re-create the iterator
-  while (bcIter != bcs.end() && bcIter.globalBC() <= maxBC && bcIter.globalBC() >= minBC) {
+  while (bcIter != bcs.end() && bcIter.globalBC() <= maxBC && (int64_t)bcIter.globalBC() >= minBC) {
     LOGF(debug, "Table id %d BC %llu", bcIter.globalIndex(), bcIter.globalBC());
     maxBCId = bcIter.globalIndex();
     ++bcIter;
